+++ src/main/java/io/beanmapper/BeanMapper.java
@@ -1,5 +1,8 @@
 package io.beanmapper;
 
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -7,6 +10,8 @@ import java.util.Set;
 
 import io.beanmapper.config.BeanMapperBuilder;
 import io.beanmapper.config.Configuration;
+import io.beanmapper.core.collections.Equalizer;
+import io.beanmapper.exceptions.BeanConversionException;
 import io.beanmapper.strategy.MapStrategyType;
 
 /**
@@ -88,6 +93,75 @@ public record BeanMapper(Configuration configuration) {
      */
     public <S, T> Set<T> map(Set<S> set, Class<T> elementInSetClass) {
         return (Set<T>) mapCollection(set, elementInSetClass);
+    }
+
+    /**
+     * Maps a Collection of elements into another, existing Collection of elements, updating the corresponding entities, and inserting elements without
+     * counterpart.
+     * @param source the source collection
+     * @param target the target collection
+     * @param <S> the class type of the elements in the  source collection, which must extend from Equalizer<ID>
+     * @param <T> the class type of the elements in the target collection, which must extend from Equalizer<ID>
+     * @return the combination of the source and target collections, where the elements of the source collection have been mapped to the target class.
+     */
+    public <S extends Equalizer, T extends Equalizer> Collection<T> map(Collection<S> source, Collection<T> target) {
+        if (target.isEmpty()) {
+            throw new RuntimeException("");
+        }
+
+
+        Set<T> set = new HashSet<>();
+        for (S entity : source) {
+            T t = null;
+            for (T otherEntity : target) {
+                if (entity.isEqual(otherEntity)) {
+                    t = map(entity, otherEntity);
+                    set.add(t);
+                    break;
+                }
+            }
+            if (t == null && !this.configuration.getOnlyPatchExistingDuringCollectionToCollection()) {
+                if (target.isEmpty()) {
+                    throw new BeanConversionException(entity.getClass(), target.getClass());
+                }
+                Class<T> clazz = (Class<T>) target.stream().findFirst().get().getClass();
+                set.add(map(entity, clazz));
+            }
+        }
+        return set;
+    }
+
+    public <S extends Equalizer, T extends Equalizer> Collection<T> map(Collection<S> source, Collection<T> target, Class<T> targetClass) {
+        // It's probably a better idea to pass the targetClass as a parameter. So this method will replace the method above this one.
+        return null;
+    }
+
+    /**
+     *Maps a Map of elements into another, existing Map of elements, updating the corresponding entities, and inserting elements without counterpart.
+     * @param source the source map
+     * @param target the target map
+     * @param targetClass the class-object representing the type of the elements in the target map.
+     * @param <K> the class of the key/ID of each element.
+     * @param <V1> the class type of the elements in the source map, which must extend from Equalizer<K>
+     * @param <V2> the class type of the elements in the target map, which must extend from Equalizer<K>
+     * @return the combination of the source and target maps, where the elements of the source map have been mapped to the target class.
+     */
+    public <K, V1 extends Equalizer, V2 extends Equalizer> Map<K, V2> map(Map<K, V1> source, Map<K, V2> target, Class<V2> targetClass) {
+        Map<K, V2> result = new HashMap<>();
+        for (var entry : source.entrySet()) {
+            boolean isMapped = false;
+            for (var targetEntry : target.entrySet()) {
+                if (entry.getValue().isEqual(targetEntry.getValue())) {
+                    result.put(targetEntry.getKey(), this.map(entry.getValue(), targetClass));
+                    isMapped = true;
+                    break;
+                }
+            }
+            if (!(isMapped || this.configuration.getOnlyPatchExistingDuringCollectionToCollection())) {
+                result.put(entry.getKey(), this.map(entry.getValue(), targetClass));
+            }
+        }
+        return result;
     }
 
     /**